+++ include/GafferSceneUI/TypeIds.h
@@ -58,6 +58,7 @@ enum TypeId
 	SetPathTypeId = 110665,
 	LightToolTypeId = 110666,
 	LightPositionToolTypeId = 110667,
+	RenderPassPathTypeId = 110668,
 
 	LastTypeId = 110700
 };