+++ Changes.md
@@ -5,6 +5,7 @@ Features
 --------
 
 - 3Delight : Added support for USD `SphereLight`, `RectLight`, `DiskLight`, `DistantLight`, `DomeLight` and `CylinderLight`.
+- RenderPassEditor : Added a new editor UI for inspecting and editing render passes.
 
 Improvements
 ------------
@@ -27,6 +28,7 @@ API
 ---
 
 - ArnoldShaderUI : Added support for `colorSpace` widget type metadata, allowing an OpenColorIO colour space to be chosen.
+- PathColumn : Added `CellData::foreground` member, to provide additional control over foreground colours in the PathListingWidget.
 
 1.3.10.0 (relative to 1.3.9.0)
 ========